+++ CHANGELOG.md
@@ -7,7 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.1.2] - 2018-11-23
-### Added
+### Changed
+- Added an optional parameter to `SQLiteDatabase#execSQL()`
 
 ## [0.1.1] - 2018-11-23
 ### Added